Add ETDumpGen_Done state and reset when in that state
Summary:
Currently when we call `get_etdump_data()` we finalize the flatbuffer and close it. In order to add entries to etdump once again after we've closed it we need to reset the etdump buffer.

In order to support this, adding a `ETDumpGen_Done` state to which we move after calling `get_etdump_data()`. When this state is detected in `create_event_block` we implicitly call `reset()` and thus enabling writing to the flatbuffer again without any user intervention.

Differential Revision: D57517927